Remove SampleOrderUtil.java
