+++ src/main/java/seedu/address/MainApp.java
@@ -28,7 +28,6 @@ import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyOrdersList;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.util.SampleDataUtil;
-import seedu.address.model.util.SampleOrderUtil;
 import seedu.address.storage.AddressBookStorage;
 import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.Storage;
@@ -103,7 +102,7 @@ public class MainApp extends Application {
             if (!ordersListOptional.isPresent()) {
                 logger.info("Data file not found. Will be starting with a sample Orders List");
             }
-            initialOrdersListData = ordersListOptional.orElseGet(SampleOrderUtil::getSampleOrdersList);
+            initialOrdersListData = ordersListOptional.orElseGet(SampleDataUtil::getSampleOrdersList);
         } catch (DataConversionException e) {
             logger.warning("Data file not in the correct format. Will be starting with an empty AddressBook");
             initialData = new AddressBook();
