+++ src/test/java/seedu/address/testutil/OrderBuilder.java
@@ -8,7 +8,7 @@ import seedu.address.model.order.Order;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
-import seedu.address.model.util.SampleOrderUtil;
+import seedu.address.model.util.SampleDataUtil;
 
 /**
  * A utility class to help with building Order objects.
@@ -53,7 +53,7 @@ public class OrderBuilder {
      * Parses the {@code food} into a {@code Set<Food>} and set it to the {@code Order} that we are building.
      */
     public OrderBuilder withFood(String ... food) {
-        this.food = SampleOrderUtil.getOrderSet(food);
+        this.food = SampleDataUtil.getOrderSet(food);
         return this;
     }
